~jakieś zmiany

+++ algorytm.h
@@ -9,9 +9,9 @@
 #include <QTextStream>
 #include "Interval.h"
 
-void returnRow(int i, int n,long  double* a, long double **input);
+void returnRow(int i, int n, double* a, double **input);
 void returnRowInterval(int i, int n, interval_arithmetic::Interval<double>* a, char ***);
-long double *GaussJordan(int n, long double result[], int *st,long double **input);
+double *GaussJordan(int n, double result[], int *st, double **input);
 void GaussJordanInterval (int n, interval_arithmetic::Interval<double> *resultInterval , int *status, char ***numbersChar);
 
 #endif // ALGORYTM_H
